Authentication added for all protected routes / FINISHED

+++ api/app.js
@@ -194,11 +194,13 @@ app.post('/lists/:listId/tasks', authenticate, (req, res) => {
     List.findOne({
         _id: req.params.listId,
         _userId: req.user_id
-    }).then((user) => {
-        if (user) {
-            // user object valide
+    }).then((list) => {
+        if (list) {
+            // list object with specified conditions found
+            // therefore the currently authenticated user can create new tasks
             return true
         }
+        // else - the list object is undefined 
         return false
     }).then(canCreateTask => {
         if (canCreateTask) {
@@ -221,17 +223,34 @@ app.post('/lists/:listId/tasks', authenticate, (req, res) => {
  * Purpose: Update an existing task
  */
 
-app.patch('/lists/:listId/tasks/:taskId', (req, res) => {
+app.patch('/lists/:listId/tasks/:taskId', authenticate, (req, res) => {
     // we want to update an existing task (specified by taskId)
-    Task.findOneAndUpdate({
-        _id: req.params.taskId,
-        _listId: req.params.listId
-    }, {
-        $set: req.body
-    }
-    ).then(() => {
-        res.send({ message: "Updated Successfully" });
-    });
+
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id,
+    }).then((list) => {
+        if (list) {
+            return true
+        }
+        return false
+    }).then((canUpdateTasks) => {
+        if (canUpdateTasks) {
+            // the currently authenticated user can update tasks 
+            Task.findOneAndUpdate({
+                _id: req.params.taskId,
+                _listId: req.params.listId
+            }, {
+                $set: req.body
+            }
+            ).then(() => {
+                res.send({ message: "Updated Successfully" });
+            });
+        } else {
+            res.sendStatus(404);
+        }
+    })
+
 
 });
 
@@ -239,14 +258,35 @@ app.patch('/lists/:listId/tasks/:taskId', (req, res) => {
  * DELETE /lists/:listId/tasks/:taskId
  * Purpose: Delete a task
  */
-app.delete('/lists/:listId/tasks/:taskId', (req, res) => {
-    // We want to delete the specified list 
-    Task.findOneAndRemove({
-        _id: req.params.taskId,
-        _listId: req.params.listId
-    }).then((removedTaskDoc) => {
-        res.send(removedTaskDoc);
+app.delete('/lists/:listId/tasks/:taskId', authenticate, (req, res) => {
+
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id,
+    }).then((list) => {
+        if (list) {
+            return true
+        }
+        return false
+    }).then((canDeleteTasks) => {
+
+        if (canDeleteTasks) {
+            Task.findOneAndRemove({
+                _id: req.params.taskId,
+                _listId: req.params.listId
+            }).then((removedTaskDoc) => {
+                res.send(removedTaskDoc);
+            });
+        }
+        else {
+            res.sendStatus(404);
+        }
+
     });
+
+
+    // We want to delete the specified list 
+
 });
 
 /* USER ROUTES */
